beta0.0

+++ data.cpp
@@ -28,19 +28,19 @@ void Data::init ()
     for (int j = i; j < _N; j++) {
       int n = j - i + 1;
 //      std::cout << "i=" << i << ", j=" << j << ", n=" << n << std::endl;
-      Eigen::MatrixXd currentMat = _contactMat.block (i, i, n, n);
-      int countIntra = (currentMat.sum () - currentMat.diagonal().sum ()) * .5;
+      Eigen::MatrixXd currentMat = _contactMat.block(i, i, n, n);
+      double countIntra = (currentMat.sum() - currentMat.diagonal().sum()) * .5;
       
-      double sum1 = _contactMat.block (0, i, i, j-i+1).sum ();
+      double sum1 = _contactMat.block(0, i, i, j-i+1).sum();
 //      std::cout << "sum1=" << sum1 << std::endl;
   
       double sum2;
       if (j + 1 < _N)
-        sum2 = _contactMat.block (i, j+1, j-i+1, _N-1-j).sum ();
+        sum2 = _contactMat.block(i, j+1, j-i+1, _N-1-j).sum();
       else
         sum2 = 0;
 //      std::cout << "sum2=" << sum2 << std::endl;
-      int countInter = sum1 + sum2;
+      double countInter = sum1 + sum2;
       _edgeCount(i, j) = countIntra;
       _edgeCount(j, i) = countInter;
     }
@@ -54,5 +54,5 @@ double Data::getVol (int s, int e)
   if (s != e)
     return 2. * _edgeCount.coeff(s, e) + _edgeCount.coeff(e, s);
   else
-    return _edgeCount.coeff (e, s);
+    return _edgeCount.coeff(e, s);
 }
